pyproject.toml
[project]
name = "odoo-bringout-new_apikey_nopassword"
version = "16.0.0"
description = "New API Key No Password - Remove password check for new API keys"
authors = [
    { name = "Ernad Husremovic", email = "hernad@bring.out.ba" }
]
dependencies = [
    "odoo-bringout-oca-ocb-base>=16.0.0",
    "requests>=2.25.1"
]
readme = "README.md"
requires-python = ">= 3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

[project.urls]
homepage = "https://github.com/bringout/0"
repository = "https://github.com/bringout/0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["new_apikey_nopassword"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.4.1",
]
